+++ src/components/Organism/CalendarDropdown/constants.ts
@@ -1,0 +1,1 @@
+export const DAY_LIST = ['일', '월', '화', '수', '목', '금', '토'];
