+++ src/components/Organism/CalendarDropdown/types.ts
@@ -1,0 +1,5 @@
+export interface CalendarDropdownProps {
+  isOpen: boolean;
+  scheduleType: 'start' | 'end';
+  selectedDate?: Date;
+}
