+++ src/components/Organism/TimePickerDropdown/index.tsx
@@ -1,44 +1,7 @@
-import React from 'react';
+import useTimePicker from '@hooks/useTimePicker';
 
-import useTimePicker, { IPickedTime } from '@hooks/useTimePicker';
-import { styled } from 'styled-components';
-import { vars } from 'token';
-
-const CalendarDropdownWrapper = styled.section<{ $isopen: boolean }>`
-  padding: 10px 32px;
-  background-color: ${vars.sementic.color.white};
-  box-shadow: 0px 0px 25px 0px rgba(0, 0, 0, 0.05);
-  display: ${(props) => (props.$isopen ? 'flex' : 'none')};
-  gap: 30px;
-  position: absolute;
-  bottom: -128px;
-  z-index: 50;
-`;
-
-const HourPicker = styled.ul`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 30px;
-  li {
-    cursor: pointer;
-  }
-`;
-
-const PickerElement = styled.li<{ $hasCurrent: boolean }>`
-  color: ${(props) =>
-    props.$hasCurrent
-      ? vars.sementic.color.black
-      : vars.sementic.color.black20};
-`;
-
-interface TimePickerDropdownProps {
-  isOpen: boolean;
-  usePickedTimeState: {
-    state: IPickedTime;
-    setState: React.Dispatch<React.SetStateAction<IPickedTime>>;
-  };
-}
+import { CalendarDropdownWrapper, HourPicker, PickerElement } from './styles';
+import { TimePickerDropdownProps } from './types';
 
 const TimePickerDropdown = ({
   isOpen,
