+++ src/components/Organism/TimePickerDropdown/types.ts
@@ -1,0 +1,9 @@
+import { IPickedTime } from '@hooks/useTimePicker';
+
+export interface TimePickerDropdownProps {
+  isOpen: boolean;
+  usePickedTimeState: {
+    state: IPickedTime;
+    setState: React.Dispatch<React.SetStateAction<IPickedTime>>;
+  };
+}
