+++ src/customTypes/calendar/calendarDate.ts
@@ -1,0 +1,6 @@
+export default interface CalendarDate {
+  id: number;
+  value: number;
+  dateType: Date;
+  isCurrentMonth: boolean;
+}
