+++ src/customTypes/calendar/index.ts
@@ -1,3 +1,4 @@
 import ICalendarDay from './ICalendarDay';
+import CalendarDate from './calendarDate';
 
-export type { ICalendarDay };
+export type { ICalendarDay, CalendarDate };
